requirements.txt
Jinja2==2.11.2
lxml==4.5.1
python-frontmatter==0.5.0
PyYAML==5.3.1
six==1.15.0
Wand==0.6.1
htmltruncate==1.0
requests==2.24.0
